feat: move success and error message functions
